+++ CHANGELOG.md
@@ -12,11 +12,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Release Notes
 - Fixed: adding/removing relays not reflected on feed filter. [#119](https://github.com/verse-pbc/issues/issues/119)
+- Added Lists view and two ways to navigate to it. [#133](https://github.com/verse-pbc/issues/issues/133)
+- Added view for editing a list's title and description. [#134](https://github.com/verse-pbc/issues/issues/134)
+- Added List detail view. [#155](https://github.com/verse-pbc/issues/issues/155)
+- Added view for managing users in a list. [#135](https://github.com/verse-pbc/issues/issues/135)
+- Added ability to delete lists. [#136](https://github.com/verse-pbc/issues/issues/136)
+- Added analytics for feed source selection and lists. [#129](https://github.com/verse-pbc/issues/issues/129)
 
 ### Internal Changes
 - Added function for creating a new list and a test verifying list editing. [#112](https://github.com/verse-pbc/issues/issues/112)
 - Localized strings on the feed filter drop-down view.
 - Disabled automatic tracking in Sentry. [#126](https://github.com/verse-pbc/issues/issues/126)
+- Improved naming of a couple list-related classes.
 - Track TestFlight vs AppStore installations in Posthog. [#130](https://github.com/verse-pbc/issues/issues/130)
 - Track breadcrumbs in Sentry for all analytics events. [#125](https://github.com/verse-pbc/issues/issues/125)